Update. All uint16_t, uint8_t to size_t. All int to int32_t. Memory check OK!

+++ src/CControl/Sources/ControlEngineering/c2d.c
@@ -11,11 +11,11 @@
  * Continuous to discrete
  * Turn A and B into discrete form
  */
-void c2d(float A[], float B[], uint8_t ADIM, uint8_t RDIM, float sampleTime) {
+void c2d(float A[], float B[], size_t ADIM, size_t RDIM, float sampleTime) {
 	size_t bytes_of_M = (ADIM + RDIM) * (ADIM + RDIM) * sizeof(float);
 	float *M = (float*)malloc(bytes_of_M);
 	memset(M, 0, bytes_of_M);
-	uint8_t i, j;
+	size_t i, j;
 	/* Create M = [A B; zeros(RDIM, ADIM) zeros(RDIM, RDIM)] */
 	for (i = 0; i < ADIM; i++) {
 		/* For A row */
